fix: conditional loading for the adobe scripts
