+++ amd/package.json
@@ -1,6 +1,6 @@
 {
   "name": "templates-common-library",
-  "version": "0.0.114",
+  "version": "0.0.115",
   "files": [
     "/*"
   ],
